Abstract test cases have Abstract prefix in their name now to make sure ant does not try to execute them.
