+++ src/test/org/apache/sqoop/cloud/s3/TestS3AvroImport.java
@@ -18,13 +18,13 @@
 
 package org.apache.sqoop.cloud.s3;
 
-import org.apache.sqoop.cloud.TestAvroImport;
+import org.apache.sqoop.cloud.AbstractTestAvroImport;
 import org.apache.sqoop.testcategories.thirdpartytest.S3Test;
 import org.junit.ClassRule;
 import org.junit.experimental.categories.Category;
 
 @Category(S3Test.class)
-public class TestS3AvroImport extends TestAvroImport {
+public class TestS3AvroImport extends AbstractTestAvroImport {
 
   @ClassRule
   public static S3CredentialsRule credentialsRule = new S3CredentialsRule();
