+++ src/test/org/apache/sqoop/cloud/s3/TestS3IncrementalAppendParquetImport.java
@@ -18,13 +18,13 @@
 
 package org.apache.sqoop.cloud.s3;
 
-import org.apache.sqoop.cloud.TestIncrementalAppendParquetImport;
+import org.apache.sqoop.cloud.AbstractTestIncrementalAppendParquetImport;
 import org.apache.sqoop.testcategories.thirdpartytest.S3Test;
 import org.junit.ClassRule;
 import org.junit.experimental.categories.Category;
 
 @Category(S3Test.class)
-public class TestS3IncrementalAppendParquetImport extends TestIncrementalAppendParquetImport {
+public class TestS3IncrementalAppendParquetImport extends AbstractTestIncrementalAppendParquetImport {
 
   @ClassRule
   public static S3CredentialsRule credentialsRule = new S3CredentialsRule();
