+++ src/test/org/apache/sqoop/cloud/s3/TestS3ParquetImport.java
@@ -18,13 +18,13 @@
 
 package org.apache.sqoop.cloud.s3;
 
-import org.apache.sqoop.cloud.TestParquetImport;
+import org.apache.sqoop.cloud.AbstractTestParquetImport;
 import org.apache.sqoop.testcategories.thirdpartytest.S3Test;
 import org.junit.ClassRule;
 import org.junit.experimental.categories.Category;
 
 @Category(S3Test.class)
-public class TestS3ParquetImport extends TestParquetImport {
+public class TestS3ParquetImport extends AbstractTestParquetImport {
 
   @ClassRule
   public static S3CredentialsRule credentialsRule = new S3CredentialsRule();
