+++ src/test/org/apache/sqoop/cloud/s3/TestS3TextImport.java
@@ -18,13 +18,13 @@
 
 package org.apache.sqoop.cloud.s3;
 
-import org.apache.sqoop.cloud.TestTextImport;
+import org.apache.sqoop.cloud.AbstractTestTextImport;
 import org.apache.sqoop.testcategories.thirdpartytest.S3Test;
 import org.junit.ClassRule;
 import org.junit.experimental.categories.Category;
 
 @Category(S3Test.class)
-public class TestS3TextImport extends TestTextImport {
+public class TestS3TextImport extends AbstractTestTextImport {
 
   @ClassRule
   public static S3CredentialsRule credentialsRule = new S3CredentialsRule();
